% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_se_cols.R
\name{add_se_cols}
\alias{add_se_cols}
\title{Add column to toptab with standard errors}
\usage{
add_se_cols(toptab, lfc.suffix = "logFC", t.suffix = "t")
}
\arguments{
\item{toptab}{Table with genes as rows and statistics as columns. Must include a column with t-statistics
e.g. named \code{"t"} or ends with \code{".t"} & a column of log fold changes e.g. named \code{"logFC"} or ends with \code{".logFC"}.
Columns with t-statistics and those with logFC's must have same prefixes.}

\item{lfc.suffix}{Suffix of logFC column(s).}

\item{t.suffix}{Suffix of t-statistic column(s).}
}
\description{
Add column to toptab with standard errors.
}
