% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zcols.R
\name{add_zcols}
\alias{add_zcols}
\title{Add column to toptab with z-stats}
\usage{
add_zcols(toptab, fit)
}
\arguments{
\item{toptab}{Table with genes as rows and statistics as columns. Must include a column with t-statistics that
is \code{"t"}, or ends with \code{".t"}.}

\item{fit}{object of class \code{\link[limma]{marraylm}}.}
}
\description{
Add column to toptab with z-stats using \code{\link[limma]{zscore}}.
}
