% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch2design.R
\name{batch2design}
\alias{batch2design}
\title{Transform factor vector into design matrix}
\usage{
batch2design(batch)
}
\arguments{
\item{batch}{Vector that is a factor or can be coerced to one.}
}
\value{
Numeric matrix.
}
\description{
Transform factor vector into design matrix, based on code from \code{\link[limma]{removeBatchEffect}}.
}
