% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_filenames.R
\name{clean_filenames}
\alias{clean_filenames}
\title{Alter names to valid filenames}
\usage{
clean_filenames(nm)
}
\arguments{
\item{nm}{Character string of names.}
}
\description{
Alter names to be valid Uniform Resource Identifier (URIs) for rJava, valid filenames in both Linux and Windows, 
and ensure they are unique. Improper characters are replaced with "_".
}
