% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pvalues.R
\name{combine_pvalues}
\alias{combine_pvalues}
\title{Combine p-values of a feature over multiple p-value columns of an object}
\usage{
combine_pvalues(mat, pv.cols = NULL)
}
\arguments{
\item{mat}{A matrix-like object with statistical columns, including some
containing p-values.}

\item{pv.cols}{the column names or column indices with p-values. If
\code{NULL}, the function searches for columns that end with \code{.p} or
\code{.pval}.}
}
\value{
p-value.
}
\description{
Combine p-values of a feature over multiple p-value columns of an object
}
\details{
This function uses the z-transform method to combine p-values, equivalently to using unweighted \code{method="z.transform"} in
\code{\link[survcomp]{combine.test}}.
}
\seealso{
\code{\link[survcomp]{combine.test}}.
}
