% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezebayes.R
\name{ezebayes}
\alias{ezebayes}
\title{Wrapper for limma's eBayes function that also allows for computing OLS statistics}
\usage{
ezebayes(fit, moderated = TRUE, trend = FALSE)
}
\arguments{
\item{fit}{Object of class \code{MArrayLM} as produced by \code{\link[limma]{lmFit}} alone (without 
\code{\link[limma]{eBayes}}).}

\item{moderated}{Logical; should \code{\link[limma]{eBayes}} be used? Otherwise an unmoderated version for 
\pkg{limma} to produce ordinary least squares statistics is used.}

\item{trend}{Logical; should an intensity-trend be allowed for the prior variance? Default is that the prior variance 
is constant.}
}
\value{
Object of \code{\link[limma]{MArrayLM-class}}.
}
\description{
Wrapper for \pkg{limma}'s \code{\link[limma]{eBayes}}, which computes empirical Bayes statistics, that allows for 
using the rest of the \pkg{limma} workflow in \pkg{ezlimma} to compute ordinary least squares statistics.
}
\details{
\code{trend} is only applicable if \code{moderated} is \code{TRUE}.
}
