% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezmediate.R
\name{ezmediate}
\alias{ezmediate}
\title{Call \code{mediate} function in \pkg{mediation} similarly to \code{hitman}}
\usage{
ezmediate(E, M, Y, covariates = NULL, sims = 10^4)
}
\arguments{
\item{E}{A numeric vector of exposures.}

\item{M}{Vector of a potential mediator.}

\item{Y}{A numeric vector of \code{length(E)} of outcomes.}

\item{covariates}{Numeric vector with one element per sample or matrix-like object with rows corresponding 
to samples and columns to covariates to be adjusted for.}

\item{sims}{Number of simulations per analyte to calculate mediation.}
}
\description{
Call \code{mediate} function in \pkg{mediation} similarly to \code{hitman}.
}
