% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezpreg.R
\name{ezpreg}
\alias{ezpreg}
\title{Test partial regression of each row of an object to a phenotype vector}
\usage{
ezpreg(object, phenotype, covar, reorder.rows = TRUE, prefix = NULL,
  adjust.method = "BH")
}
\arguments{
\item{object}{A matrix-like data object containing log-ratios or 
log-expression values, with rows corresponding to 
genes and columns to samples.}

\item{phenotype}{Vector of phenotypes of the samples. Should be same length as
\code{ncol(object)}. If the vector is named, names should match 
\code{colnames(object)}.}

\item{covar}{A matrix or dataframe with one or more covariates as columns.}

\item{reorder.rows}{logical, should rows be reordered by p-value or be left in 
the same order as \code{object}?}

\item{prefix}{character string to add to beginning of column names.}

\item{adjust.method}{method used to adjust the p-values for multiple testing.}
}
\value{
Dataframe.
}
\description{
Test partial regression of each row of an object to a phenotype vector given covariates in a linear 
regression of \code{phenotype ~ 1 + x + covar} for \code{x} being each row of \code{object}.
}
