% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eztoptab.R
\name{eztoptab}
\alias{eztoptab}
\title{Wrapper for limma topTable function}
\usage{
eztoptab(fit, cols = c("P.Value", "adj.P.Val", "logFC"),
  adjust.method = "BH", prefix = "", coef = NULL)
}
\arguments{
\item{fit}{should be an object of class \code{MArrayLM} as produced by 
\code{lmFit} and \code{eBayes}.}

\item{cols}{columns of \code{topTable} output the user would like in the 
output, although the names are changed. If \code{logFC} is specified, 
\code{FC} will also be given.}

\item{adjust.method}{method used to adjust the p-values for multiple testing.}

\item{prefix}{character string to add to beginning of column names.}

\item{coef}{column number or column name specifying which coefficient or 
contrast of the linear model is of interest. If \code{length(coef)>1}, an F-test will be performed,
and \code{logFC} & \code{FC} will not be returned.}
}
\value{
Dataframe.
}
\description{
Wrapper for limma topTable function that subsets and changes colnames
}
\details{
See \code{\link[limma]{topTable}} for more details, as this fuction is a wrapper for that one.
}
