% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc2ratio.R
\name{fc2ratio}
\alias{fc2ratio}
\title{Transform fold-change (FC) into a ratio}
\usage{
fc2ratio(FC)
}
\arguments{
\item{FC}{Numeric vector of fold-changes.}
}
\value{
Vector of corresponding ratios.
}
\description{
Transform fold-change (FC) into a ratio. It's assumed \pkg{ezlimma} tranformed a ratio less than 1 into
a negative FC, e.g. a ratio of 0.5 gave a FC of -2. This function reverses that operation.
}
