% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmat2gmt.R
\name{gmat2gmt}
\alias{gmat2gmt}
\title{Transform analyte-by-pathway matrix into list compatible with Gene Matrix Transposed (GMT) file format}
\usage{
gmat2gmt(gmat)
}
\arguments{
\item{gmat}{A numeric Matrix of features (proteins and/or metabolites) as rows and pathways as columns, indicating 
pathway membership with positive elements.}
}
\value{
A list of analyte sets. Each element has a set \code{name}, \code{description}, and \code{genes}.
}
\description{
Transform analyte-by-pathway matrix into list compatible with Gene Matrix Transposed (GMT) file format, described at the \href{https://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#GMT:_Gene_Matrix_Transposed_file_format_.28.2A.gmt.29}{Broad Institute}.
}
