% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grep_cols.R
\name{grep_cols}
\alias{grep_cols}
\title{Find p-value or stat columns in table}
\usage{
grep_cols(tab, p.cols = NULL, stat.cols = NULL)
}
\arguments{
\item{tab}{Matrix-like object with statistical columns, some containing p-values.}

\item{p.cols}{Indices or \code{\link{regexp}} with column names or column names suffix of numeric p-value columns.}

\item{stat.cols}{Indices or \code{\link{regexp}} with column names or column names suffix with numeric signed 
statistics, or with \code{"Up", "Down"} values. These should match \code{p.cols}.}
}
\value{
Vector of column names.
}
\description{
Find p-value or stat columns in table and check their validity.
}
\details{
Exactly one of \code{p.cols} or \code{stat.cols} should not be \code{NULL}.
}
