% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_means.R
\name{grp_means}
\alias{grp_means}
\title{Calculate group means per feature}
\usage{
grp_means(object, grp, args.lst = list())
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to 
features (e.g. genes) and columns to samples. Must have row names that are non-duplicated and non-empty.}

\item{grp}{Vector of sample groups. These must be valid variable names in R and the same length as 
\code{ncol(object)}.}

\item{args.lst}{List of arguments to be used in \code{do.call(limma::lmFit, args.list)}. It is only checked
whether \code{is.null(args.lst$ndups)} and \code{is.null(args.lst$spacing)}, which could be true or false.}
}
\value{
Matrix of features (e.g. genes) by groups of means of object with column names ending in \code{.avg}.
}
\description{
Calculate group means per feature (e.g. per gene), accounting for possible intra-sample duplication.
}
