% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitman.R
\name{hitman}
\alias{hitman}
\title{High-throughput mediation analysis}
\usage{
hitman(E, M, Y, covariates = NULL, verbose = FALSE)
}
\arguments{
\item{E}{A vector representing exposure; can be a character with groups or numeric.}

\item{M}{A numeric matrix-like data object with one row per feature and one column per sample representing the mediator.}

\item{Y}{A numeric vector of \code{length(E)} of outcomes.}

\item{covariates}{Numeric vector or matrix of covariates.}

\item{verbose}{Logical indicating if messages should be reported.}
}
\description{
High-throughput mediation analysis to test if rows of \code{M} mediate the effect of exposure \code{E} on outcome
\code{Y}.
}
