% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_signif_mediation.R
\name{joint_signif_mediation}
\alias{joint_signif_mediation}
\title{Joint Significant mediation test}
\usage{
joint_signif_mediation(E, M, Y, covariates = NULL)
}
\arguments{
\item{E}{A numeric vector of exposures.}

\item{M}{A numeric vector  of mediators with one element per sample.}

\item{Y}{A numeric vector of \code{length(E)} of outcomes.}

\item{covariates}{Numeric vector with one element per sample or matrix-like object with rows corresponding 
to samples and columns to covariates to be adjusted for.}
}
\value{
p-value.
}
\description{
Joint Significant mediation test.
}
\details{
\code{E} and \code{Y} cannot have \code{NA}s.
}
\seealso{
\url{https://github.com/cedricbatailler/JSmediation}.
}
