% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_contrasts.R
\name{limma_contrasts}
\alias{limma_contrasts}
\title{Apply limma's lmFit, contrasts.fit, & eBayes to one or more contrasts and return a table}
\usage{
limma_contrasts(
  object,
  grp = NULL,
  contrast.v,
  design = NULL,
  weights = NA,
  trend = FALSE,
  ndups = NULL,
  spacing = NULL,
  block = NULL,
  correlation = NULL,
  adjust.method = "BH",
  add.means = !is.null(grp),
  treat.lfc = NULL,
  moderated = TRUE,
  check.names = TRUE,
  cols = c("P.Value", "adj.P.Val", "logFC")
)
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to 
features (e.g. genes) and columns to samples. Must have row names that are non-duplicated and non-empty.}

\item{grp}{Vector of sample groups. These must be valid variable names in R and the same length as 
\code{ncol(object)}.}

\item{contrast.v}{Named vector of contrasts, passed to \code{\link[limma]{makeContrasts}}.}

\item{design}{Design matrix of the experiment, with rows corresponding to samples and columns to coefficients to be 
estimated.}

\item{weights}{Non-negative observation weights. Can be a numeric matrix of individual weights of same size as the 
\code{object}, or a numeric vector of sample weights with length \code{ncol(object)}, or a numeric vector of gene 
weights with length equal to \code{nrow(object)}. Set to \code{NULL} to ignore \code{object$weights}. 
\code{weights=NA} (with length one) doesn't pass weights to \code{limma}.}

\item{trend}{Logical; should an intensity-trend be allowed for the prior variance? Default is that the prior variance 
is constant.}

\item{ndups}{Positive integer giving the number of times each distinct probe is measured in each sample. See Details.}

\item{spacing}{Positive integer giving the spacing between duplicate occurrences of the same probe, with 
\code{spacing=1} for consecutive rows.  See Details.}

\item{block}{Vector specifying a blocking variable on the samples. Has length = \code{ncol(object)}. 
Must be \code{NULL} if \code{ndups > 1}.}

\item{correlation}{Inter-duplicate or inter-technical replicate correlation. Must be given if 
\code{ndups>1} or \code{!is.null(block)}.}

\item{adjust.method}{Method used to adjust the p-values for multiple testing. Options, in increasing conservatism, 
include \code{"none"}, \code{"BH"}, \code{"BY"}, and \code{"holm"}. See \code{\link[stats]{p.adjust}} for the complete
list of options. A \code{NULL} value will result in the default adjustment method, which is \code{"BH"}.}

\item{add.means}{Logical indicating if (unweighted) group means per row should be added to the output.}

\item{treat.lfc}{Vector of logFC passed to \code{\link[limma:lmFit]{treat}} \code{lfc}. It is recycled as needed to match
rows of \code{object}. If given, \code{length(contrast.v)} must be 1. McCarthy & Smyth suggest a 10 percent fold-change,
which is \code{treat.lfc=log2(1.1)}. Limma Treat tests for logFC's outside the interval [-treat.lfc, treat.lfc].}

\item{moderated}{Logical; should \code{\link[limma]{eBayes}} be used? Otherwise an unmoderated version for 
\pkg{limma} to produce ordinary least squares statistics is used.}

\item{check.names}{Logical; should \code{names(grp)==rownames(object)} be checked? Ignored if \code{is.null(design)}
and \code{add.means} is \code{FALSE}.}

\item{cols}{Columns of \code{topTable} output to include. Possibilities include 
\code{"logFC", "SE", "AveExpr", "z", "t", "P.Value", "adj.P.Val", "CI.L", "CI.R", "B"}. Some of these column names are then changed here. 
If \code{logFC} is specified, \code{FC} will automatically also be given.}
}
\value{
Data frame.
}
\description{
Apply \pkg{limma}'s \code{lmFit}, \code{contrasts.fit}, & \code{eBayes} to one or more contrasts, and return
a table. See examples in vignette.
}
\details{
If \code{design} is \code{NULL} and \code{grp} is given, design will be calculated as 
\code{model.matrix(~0+grp)}. However, \code{grp} isn't needed if \code{design} is provided & \code{add.means} 
is \code{FALSE}.

The defaults of arguments \code{ndups} and \code{spacing} are set to \code{NULL}, 
which allows these arguments to be overridden by the elements \code{object$printer$ndups} and
\code{object$printer$spacing}, respectively, if these exist. Whereas, if an element does not exist,
the corresponding argument is treated as being its default in \code{\link[limma]{lmFit}}, 
i.e. \code{ndups=1} or \code{spacing=1}. 
If either of these arguments are specified, they would override the 
respective element of \code{object$printer}, if the element existed.

When \code{moderated} is FALSE, an error is generated if \code{!is.null(treat.lfc)} or \code{trend} is TRUE.

Cols \code{CI.L} and \code{CI.R} give the 95% confidence interval of the \code{logFC}. \code{SE} gives the empirical Bayes standard error.
}
\references{
McCarthy DJ & Smyth GK (2009). Testing significance relative to a fold-change threshold is a TREAT. 
Bioinformatics 25, 765-771.
}
\seealso{
\code{\link[limma]{lmFit}}; \code{\link[limma]{eBayes}}; \code{\link{limma_cor}}.
}
