% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_dep.R
\name{limma_dep}
\alias{limma_dep}
\title{Test (conditional) dependence of two variables}
\usage{
limma_dep(object, Y, covariates = NULL, prefix = NULL)
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to 
features (e.g. genes) and columns to samples. Must have rownames that are non-duplicated and non-empty.}

\item{Y}{A numeric vector or matrix.}

\item{covariates}{A numeric matrix-like data object with rows corresponding to samples and columns to covariates
to be adjusted for.}

\item{prefix}{Character string to add to beginning of column names.}
}
\value{
Data frame with \code{t-statistic} or \code{F-statistic} and \code{p-value} per row of \code{object}.
}
\description{
Test (conditional) dependence of \code{Y} with rows of \code{object} via \code{\link[ezlimma]{limmaF}}. Let a row of
\code{object} be \eqn{o}, then we test the coefficient(s) of \code{Y} in the model \eqn{o = 1 + Y + covariates}. All
inputs should be numeric, unlike \code{\link[ezlimma]{limma_contrasts}}.
}
