% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_mediation.R
\name{limma_mediation}
\alias{limma_mediation}
\title{Mediation analysis}
\usage{
limma_mediation(E, M, Y, covar = NULL)
}
\arguments{
\item{E}{A vector representing exposure; can be a character with groups or numeric}

\item{M}{A numeric matrix-like data object with one row per feature and one column per sample representing the mediator}

\item{Y}{A numeric vector of \code{length(E)} of outcomes}

\item{covar}{A matrix-like data object with rows corresponding to samples and columns to covariates. If \code{covar}
has rownames, they should match \code{names(E)}.}
}
\description{
Mediation analysis to test if rows of \code{A} mediate the effect of exposure \code{E} on \code{B}.
}
