% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_pcor.R
\name{limma_pcor}
\alias{limma_pcor}
\title{Test partial correlation of each row of an object to a phenotype vector}
\usage{
limma_pcor(object, phenotype, batch = NULL, covariates = NULL,
  reorder.rows = TRUE, prefix = NULL, adjust.method = "BH")
}
\arguments{
\item{object}{A matrix-like data object containing log-ratios or 
log-expression values, with rows corresponding to 
genes and columns to samples.}

\item{phenotype}{Vector of phenotypes of the samples. Should be same length as
\code{ncol(object)}. If the vector is named, names should match 
\code{colnames(object)}.}

\item{batch}{Batch vector passed to \code{\link[limma]{removeBatchEffect}}, which coerces it to factor.}

\item{covariates}{Numeric matrix with one or more covariates as columns passed to 
\code{\link[limma]{removeBatchEffect}}.}

\item{reorder.rows}{logical, should rows be reordered by p-value or be left in 
the same order as \code{object}?}

\item{prefix}{character string to add to beginning of column names.}

\item{adjust.method}{method used to adjust the p-values for multiple testing.}
}
\value{
Dataframe.
}
\description{
Test partial correlation of each row of an object to a phenotype vector given covariates.
}
\details{
\code{\link[limma]{removeBatchEffect}} treats \code{batch} differently form \code{covariates}, even if 
\code{covariates=as.numeric(as.factor(batch))}.
}
