% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_tost.R
\name{limma_tost}
\alias{limma_tost}
\title{Two one-sided test procedure with \pkg{limma} for statistical equivalence test}
\usage{
limma_tost(object, grp = NULL, contrast.v, tost.lfc, design = NULL,
  weights = NA, trend = FALSE, block = NULL, correlation = NULL,
  adjust.method = "BH", add.means = !is.null(grp),
  check.names = TRUE)
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to 
features (e.g. genes) and columns to samples. Must have rownames that are non-duplicated and non-empty.}

\item{grp}{Vector of sample groups. These must be valid variable names in R and the same length as 
\code{ncol(object)}.}

\item{contrast.v}{Named vector of contrasts, passed to \code{\link[limma]{makeContrasts}}.}

\item{tost.lfc}{Numeric; smallest logFC that could be of interest.}

\item{design}{Design matrix of the experiment, with rows corresponding to samples and columns to coefficients to be 
estimated.}

\item{weights}{Non-negative observation weights. Can be a numeric matrix of individual weights of same size as the 
\code{object}, or a numeric vector of sample weights with length \code{ncol(object)}, or a numeric vector of gene 
weights with length equal to\code{nrow(object)}. Set to \code{NULL} to ignore \code{object$weights}. 
\code{weights=NA} (with length one) doesn't pass weights to \code{limma}.}

\item{trend}{Logical; should an intensity-trend be allowed for the prior variance? Default is that the prior variance 
is constant.}

\item{block}{Vector specifying a blocking variable on the samples. Has length = \code{ncol(object)}.}

\item{correlation}{Inter-duplicate or inter-technical replicate correlation.}

\item{adjust.method}{Method used to adjust the p-values for multiple testing. Options, in increasing conservatism, 
include \code{"none"}, \code{"BH"}, \code{"BY"}, and \code{"holm"}. See \code{\link[stats]{p.adjust}} for the complete
list of options. A \code{NULL} value will result in the default adjustment method, which is \code{"BH"}.}

\item{add.means}{Logical indicating if (unweighted) group means per row should be added to the output.}

\item{check.names}{Logical; should \code{names(grp)==rownames(object)} be checked? Ignored if \code{is.null(design)}
and \code{add.means} is \code{FALSE}.}
}
\description{
Two one-sided test procedure with \pkg{limma} for statistical equivalence test of one contrast.
}
\details{
\code{length(contrast.v)} must be 1. This function is based on \code{\link{limma_contrasts}}, so see there
for more details.
}
\seealso{
\code{\link{limma_contrasts}}.
}
