% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaF.R
\name{limmaF}
\alias{limmaF}
\title{F-test of each row of object using design matrix}
\usage{
limmaF(object, design, coef = colnames(design), prefix = NULL,
  weights = NULL, trend = FALSE, block = NULL, correlation = NULL,
  adjust.method = "BH", reorder.rows = TRUE, cols = c("F",
  "P.Value"))
}
\arguments{
\item{object}{Matrix-like data object containing log-ratios or log-expression values, with rows corresponding to 
features (e.g. genes) and columns to samples. Must have rownames that are non-duplicated and non-empty.}

\item{design}{Design matrix of the experiment, with rows corresponding to samples and columns to coefficients to be 
estimated.}

\item{coef}{Column index or column name of the linear model to test, passed to \code{\link{eztoptab}}.}

\item{prefix}{Character string to add to beginning of column names.}

\item{weights}{Non-negative observation weights. Can be a numeric matrix of individual weights of same size as the 
\code{object}, or a numeric vector of sample weights with length \code{ncol(object)}, or a numeric vector of gene 
weights with length equal to\code{nrow(object)}. Set to \code{NULL} to ignore \code{object$weights}. 
\code{weights=NA} (with length one) doesn't pass weights to \code{limma}.}

\item{trend}{Logical, should an intensity-trend be allowed for the prior variance? Default is that the prior variance 
is constant.}

\item{block}{Vector specifying a blocking variable on the samples. Has length = \code{ncol(object)}.}

\item{correlation}{Inter-duplicate or inter-technical replicate correlation.}

\item{adjust.method}{Method used to adjust the p-values for multiple testing. Options, in increasing conservatism, 
include \code{"none"}, \code{"BH"}, \code{"BY"}, and \code{"holm"}. See \code{\link[stats]{p.adjust}} for the complete
list of options. A \code{NULL} value will result in the default adjustment method, which is \code{"BH"}.}

\item{reorder.rows}{Logical, should rows be reordered by p-value?}

\item{cols}{Columns of \code{topTable} output to include. Possibilities include \code{"logFC", "AveExpr", "t", "P.Value", 
"adj.P.Val", "B"}. Some of these column names are then changed here. If \code{logFC} is specified, \code{FC} will 
automatically also be given.}
}
\value{
Data frame.
}
\description{
F-test of each row of object using design matrix for selected coefficients. If there's only one selected coefficient, 
a t-test is applied.
}
\examples{
object <- matrix(rnorm(60), nrow=10, ncol=6)
grp <- rep(c("A", "B", "ctrl"), each=2)
covar <- cbind(covar1=1:6, covar2=rnorm(6))
design <- model.matrix(~0+grp+covar)
lf <- limmaF(object=object, design=design, coef=colnames(design)[1:3])
}
\seealso{
\code{\link[limma]{lmFit}} and \code{\link[limma]{eBayes}}.
}
