% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfc2fc.R
\name{logfc2fc}
\alias{logfc2fc}
\title{Transform a log2 fold-change to a fold-change}
\usage{
logfc2fc(logFC)
}
\arguments{
\item{logFC}{A numeric log2 fold-change.}
}
\value{
The fold-change.
}
\description{
Transform a possibly negative log2 fold-change to a fold-change.
}
\details{
This function assumes that log fold-changes less than 1 were converted 
to their negative multiplicative reciprocal, e.g. lfc = 0.5 -> lfc = -2, as is
commonly done in biology.
}
\examples{
logfc2fc(2) #4
logfc2fc(-2) #-4
}
