% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_glist.R
\name{map_glist}
\alias{map_glist}
\title{Gets the mapped IDs (rownames) of features (e.g. genes) for a gene set list of a particular type (e.g. gene symbols)}
\usage{
map_glist(G, annot, sep.str = " /// ", symbol.col = "Gene.Symbol")
}
\arguments{
\item{G}{Gene set list as returned from \code{\link{read_gmt}}.}

\item{annot}{annotation for the features that has a column of the same type as in gene set list *G*.}

\item{sep.str}{strings that separates symbols if there are multiple symbols for a feature.}

\item{symbol.col}{column name or index for the symbol column in *annot*.}
}
\value{
Gene set list *G* with the symbols replaced by the mapped IDs.
}
\description{
Gets the mapped IDs (rownames) of features (e.g. genes) for a gene set list of a particular type (e.g. gene symbols)
}
