% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_hitman_pvalues.R
\name{modify_hitman_pvalues}
\alias{modify_hitman_pvalues}
\title{Modify p-values of hitman or lotman}
\usage{
modify_hitman_pvalues(tab, overall.sign, stat.cols = c("EM.t", "MY.t"),
  p.cols = c("EM.p", "MY.p"))
}
\arguments{
\item{tab}{Numeric matrix. Must have "EM" and "MY" stat and ".p" columns.}

\item{overall.sign}{Sign of overall effect from exposure to outcome, one of 1 or -1.}

\item{stat.cols}{Vector of length 2 with column names or indices of signed statistics.}

\item{p.cols}{Vector of length 2 with column names or indices of p-values.}
}
\value{
Matrix with p-value columns modified.
}
\description{
Modify p-values of hitman or lotman based on sidedness of both tests.
}
