% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mroast_two2one_tailed.R
\name{mroast_two2one_tailed}
\alias{mroast_two2one_tailed}
\title{Convert 2-tailed into 1-tailed p-values from mroast}
\usage{
mroast_two2one_tailed(tab, pv.col = "PValue", dir.col = "Direction",
  direction = "Up", nrot = 9999)
}
\arguments{
\item{tab}{table with statistics from \code{mroast}.}

\item{pv.col}{name or index of p-value column.}

\item{dir.col}{name or index of column giving direction gene set has changed.}

\item{direction}{direction of gene set change. Can be \code{"Up"} or \code{"Down"}.}
}
\value{
Vector of p-values.
}
\description{
Convert 2-tailed into 1-tailed p-values from \code{limma} function \code{mroast}.
}
