% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_cor.R
\name{multi_cor}
\alias{multi_cor}
\title{Test correlation of each row of an object to each column of pheno.mat}
\usage{
multi_cor(object, pheno.mat, method = c("pearson", "spearman", "kendall",
  "limma"), reorder.rows = TRUE, prefix = "", adjust.method = "BH",
  limma.cols = c("AveExpr", "P.Value", "adj.P.Val", "logFC"))
}
\arguments{
\item{object}{A matrix-like data object containing log-ratios or 
log-expression values, with rows corresponding to features (eg genes) and 
columns to samples.}

\item{pheno.mat}{matrix-like data object of phenotypes of the samples, with each column one 
phenotype vector. Length and names of rows of \code{pheno.mat} should 
correspond to columns of \code{object}.}

\item{method}{a character string indicating which association is to be used 
for the test. One of \code{"pearson"}, \code{"spearman"}, \code{"kendall"}, 
from \code{\link[stats]{cor.test}} or \code{"limma"} for 
\code{\link{limma_cor}}.}

\item{reorder.rows}{logical, should rows be reordered by F-statistic from 
\code{\link[limma]{toptable}} or be left in the same order as 
\code{object}?}

\item{prefix}{character string to add to beginning of column names.}

\item{adjust.method}{method used to adjust the p-values for multiple testing.}

\item{limma.cols}{if \code{method="limma"}, this specifies \code{cols} from 
\code{\link{limma_cor}}. Ignored without a warning if \code{method} not 
\code{"limma"}.}
}
\value{
Dataframe with several statistical columns corresponding to each
 phenotype and one row per feature.
}
\description{
Test correlation of each row of an object to each column of pheno.mat using 
one of Pearson's, Kendall's, or Spearman's correlation methods, or limma 
regression in \code{\link{limma_cor}}.
}
