% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_genesettest.R
\name{multi_genesettest}
\alias{multi_genesettest}
\title{A wrapper function for \code{\link[limma]{geneSetTest}} with output to Excel}
\usage{
multi_genesettest(gstats, G, feat.tab, name = NA,
  alternative = "mixed", type = "auto", adjust.method = "BH",
  min.nfeats = 3, max.nfeats = 1000, ranks.only = TRUE,
  nsim = 9999)
}
\arguments{
\item{gstats}{A named vector, matrix or data.frame of genewise statistic (e.g. z-scores, t-statistics) by which genes can be ranked. 
The names or rownames should be the same as the rownames of  *feat.tab*}

\item{G}{Gene set list as returned from \code{\link{read_gmt}}.}

\item{feat.tab}{Table of feature (e.g. gene) statistics that the Excel table can link to.}

\item{name}{Name for the folder and Excel file that get written. Set to \code{NA} to avoid writing output.}

\item{alternative}{Alternative hypothesis; must be one of \code{"two.sided"}, \code{"greater"} or \code{"less"}. 
\code{"greater"} corresponds to positive association, \code{"less"} to negative association.}

\item{type}{character string specifying whether the statistics are signed (t-like, \code{"t"}) or unsigned (F-like, \code{"f"}) or whether the function should make an educated guess (\code{"auto"}).
  If the statistic is unsigned, then it assume that larger statistics are more significant.}

\item{adjust.method}{Method used to adjust the p-values for multiple testing. Only for \code{mroast}.}

\item{min.nfeats}{Minimum number of features (e.g. genes) needed in a gene set for testing.}

\item{max.nfeats}{Maximum number of features (e.g. genes) needed in a gene set for testing.}

\item{ranks.only}{logical, if \code{TRUE} only the ranks of the \code{statistics} are used.}

\item{nsim}{number of random samples to take in computing the p-value.
  Not used if \code{ranks.only=TRUE}.}
}
\value{
Data frame of gene set statistics.
}
\description{
Test whether a set of genes is highly ranked. It returns a data frame with statistics per gene set, and writes this to an Excel file. 
The Excel file links to CSV files, which contain statistics per gene set.
}
\details{
Pathway (i.e. gene set) names are altered to be valid filenames in Windows and Linux. Numeric columns are
rounded to 3 significant figures.
}
