% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTopTab.R
\name{multiTopTab}
\alias{multiTopTab}
\title{Combine one or more toptables extracted from linear model fit}
\usage{
multiTopTab(fit, cols = c("P.Value", "adj.P.Val", "logFC"),
  adjust.method = "BH")
}
\arguments{
\item{fit}{should be an object of class \code{MArrayLM} as produced by 
\code{lmFit} and \code{eBayes}.}

\item{cols}{columns of \code{topTable} output the user would like in the 
output, although the names are changed. If \code{logFC} is specified, 
\code{FC} will also be given.}

\item{adjust.method}{method used to adjust the p-values for multiple testing.}
}
\value{
Dataframe
}
\description{
Combine one or more toptables extracted from \code{fit}.
}
\details{
This function calls \code{\link{eztoptab}}. See also 
\code{\link{[limma]{topTable}}} for more details on many of these, as this 
fuction is a wrapper for that one.
}
