% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_changed.R
\name{prop_changed}
\alias{prop_changed}
\title{Calculate proportion of changed features per pathway}
\usage{
prop_changed(feat.tab, index, stat.col = 1, p.col = 2)
}
\arguments{
\item{feat.tab}{Table of feature (e.g. gene) statistics.}

\item{index}{List returned from \code{g_index}.}

\item{stat.col}{Name or index of column with signed statistics.}

\item{p.col}{Name or index of p-value column.}
}
\description{
Calculate proportion of features up & down at p=0.05 per pathway.
}
