% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_linked_xl.R
\name{read_linked_xl}
\alias{read_linked_xl}
\title{Read Excel XLSX file with links to CSVs}
\usage{
read_linked_xl(file.xlsx)
}
\arguments{
\item{file.xlsx}{Path of XLSX file.}
}
\value{
Data frame.
}
\description{
Read Excel XLSX file with links to CSVs written by \code{write_linked_xl}.
}
