% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescue_filenames.R
\name{rescue_filenames}
\alias{rescue_filenames}
\title{Alter names to valid filenames}
\usage{
rescue_filenames(nm)
}
\arguments{
\item{nm}{Character string of names.}
}
\description{
Alter names to be valid Uniform Resource Identifier (URIs) for rJava, valid filenames in both Linux and Windows, 
and ensure they are unique. Improper characters are replaced with "_".
}
