% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roast_two2one_tailed.R
\name{roast_two2one_tailed}
\alias{roast_two2one_tailed}
\title{Convert roast two-tailed results to one-tailed}
\usage{
roast_two2one_tailed(roast.res, fun, alternative, nrot, adjust.method)
}
\arguments{
\item{roast.res}{Result from \code{mroast} or \code{fry}.}

\item{fun}{Function to use, either \code{fry} or \code{mroast}.}

\item{alternative}{Alternative hypothesis; must be one of \code{"two.sided"}, \code{"greater"} or \code{"less"}. 
\code{"greater"} corresponds to positive association, \code{"less"} to negative association.}

\item{nrot}{Number of rotations used to estimate the p-values for \code{mroast}.}

\item{adjust.method}{Method used to adjust the p-values for multiple testing. Options, in increasing conservatism, 
include \code{"none"}, \code{"BH"}, \code{"BY"}, and \code{"holm"}. See \code{\link[stats]{p.adjust}} for the complete
list of options. A \code{NULL} value will result in the default adjustment method, which is \code{"BH"}.}
}
\value{
Modified result from \code{mroast} or \code{fry}.
}
\description{
Convert \code{mroast} or \code{fry} two-tailed results to one-tailed and remove 
non-directional, termed "Mixed", statistics.
}
