% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_barfield.R
\name{sim_barfield}
\alias{sim_barfield}
\title{Simulate & test mediation methods as in Barfield et al. (2017)}
\usage{
sim_barfield(med.fcn, b1t2.v = c(0, 0.14, 0.39), alpha = 0.05,
  nsamp = 50, nsim = 10^4, ngene = 0, seed = 1, verbose = TRUE,
  ...)
}
\arguments{
\item{med.fcn}{Mediation function to test. Must accept parameters and output matrix with mediation p-value 
in column \code{"EMY.p"}, like \code{\link{hitman}} or \code{\link{lotman}}. Its warnings are suppressed.}

\item{b1t2.v}{Numeric vector of values that both theta2 (\code{"t2"}) and beta1 (\code{"b1"}) take.}

\item{alpha}{Alpha level.}

\item{nsamp}{Number of samples.}

\item{nsim}{Number of simulations.}

\item{ngene}{Number of genes other than that of primary interest to simulate.}

\item{seed}{Random seed.}

\item{verbose}{Logical; should the number of simulations be printed every 100 simulations?}

\item{...}{Sent to \code{med.fcn}}
}
\value{
Matrix with proportion of significant calls for every combination of \code{t2} and \code{b1}.
}
\description{
Simulate & test mediation methods as in Barfield et al. (2017). If multiple genes are simulated, only 1st 
is tested.
}
\references{
Barfield R, Shen J, Just AC, Vokonas PS, Schwartz J, Baccarelli AA, VanderWeele TJ, Lin X. 
Testing for the indirect effect under the null for genome-wide mediation analyses. Genet Epidemiol. 
2017 Dec;41(8):824-833.
}
