% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_fisher.R
\name{sim_fisher}
\alias{sim_fisher}
\title{Apply \code{fisher_enrichment_test} to simulated data}
\usage{
sim_fisher(G, feat.tab, grp, effect.v = c(0, 0.2), alpha = 0.2,
  nsim = 99, test.ft = TRUE, seed = 1, verbose = FALSE)
}
\arguments{
\item{G}{Gene set list as returned from \code{\link{read_gmt}}.}

\item{feat.tab}{Table of feature (e.g. gene) statistics that the Excel table can link to.}

\item{grp}{Vector of sample groups. These must be valid variable names in R and the same length as 
\code{ncol(object)}.}

\item{effect.v}{Numeric vector of log fold-changes or percent of phentotypes to add.}

\item{alpha}{Alpha level.}

\item{nsim}{Number of simulations.}

\item{test.ft}{Compare to \code{\link[stats]{fisher.test}}.}

\item{seed}{Integer seed to set for reproducility if \code{fun="mroast"}, since \code{mroast} uses random 
simulations. Ignored if \code{fun="fry"}.}

\item{verbose}{Logical; should the number of simulations be printed every 100 simulations?}
}
\description{
Apply \code{fisher_enrichment_test} to simulated data.
}
