% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_fisher_hitman.R
\name{sim_fisher_hitman}
\alias{sim_fisher_hitman}
\title{Apply \code{fisher_enrichment_test} to simulated data analyzed by \code{hitman}}
\usage{
sim_fisher_hitman(G, E, effect.v = c(0, 0.2), alpha = 0.05,
  nsim = 10^3, seed = 1, verbose = TRUE)
}
\arguments{
\item{G}{Gene set list as returned from \code{\link{read_gmt}}.}

\item{E}{A numeric vector of exposures.}

\item{effect.v}{Numeric vector of log fold-changes or percent of phentotypes to add.}

\item{alpha}{Alpha level.}

\item{nsim}{Number of simulations.}

\item{seed}{Integer seed to set for reproducility if \code{fun="mroast"}, since \code{mroast} uses random 
simulations. Ignored if \code{fun="fry"}.}

\item{verbose}{Logical; should the number of simulations be printed every 100 simulations?}
}
\description{
Apply \code{fisher_enrichment_test} to simulated data analyzed by \code{hitman}.
}
