% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_xl.R
\name{top_xl}
\alias{top_xl}
\title{Make XL data frame for \code{writexl}}
\usage{
top_xl(pwy.tab, n.toptabs = Inf)
}
\arguments{
\item{pwy.tab}{A table of pathway results.}

\item{n.toptabs}{Number of gene set toptables to write to CSV and link to from Excel.}
}
\value{
A list with 2 elements
\describe{
  \item{\code{xl}}{data frame with xl links}
  \item{\code{pwy.csv.nms}}{vector of pathway names to write to csv files}
}
}
\description{
Make XL data frame for \code{writexl} using top pathways table.
}
