% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two2one_tailed.R
\name{two2one_tailed}
\alias{two2one_tailed}
\title{Convert 2-tailed into 1-tailed p-values}
\usage{
two2one_tailed(tab, stat.col = 1, p.col = 2,
  alternative = c("greater", "less"))
}
\arguments{
\item{tab}{Table with statistics from \code{fry}.}

\item{stat.col}{Name or index of column with signed statistics.}

\item{p.col}{Name or index of p-value column.}

\item{alternative}{Vector of direction of change: either \code{"greater"} or \code{"less"}.}
}
\value{
p-values.
}
\description{
Convert 2-tailed into 1-tailed p-values from matrix. Alternative "greater" corresponds to positive association, 
"less" to negative association.
}
