% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gmt.R
\name{write_gmt}
\alias{write_gmt}
\title{Write gene set file with GMT extension from a list}
\usage{
write_gmt(gmt.lst, file.gmt)
}
\arguments{
\item{gmt.lst}{List of gene sets. Each element has a gene set \code{name}, \code{description}, and \code{genes}.}

\item{file.gmt}{Name of the GMT file to write to. \code{.gmt} is appended.}
}
\value{
Invisibly, \code{TRUE}.
}
\description{
Write Gene Matrix Transposed (gmt) file from a list. The first column of
the GMT are gene set names, the second are descriptions, then are genes.
}
\details{
This function was adapted from \code{gmtlist2file} in the \pkg{cogena} package.
}
