% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_linked_xlsx.R
\name{write_linked_xlsx}
\alias{write_linked_xlsx}
\title{Write Excel xlsx file with links to CSVs}
\usage{
write_linked_xlsx(name, fun, res, index, stats.tab, n.toptabs)
}
\arguments{
\item{name}{a name for the folder and Excel file that get written.}

\item{fun}{function to use, either \code{fry} or \code{mroast}.}

\item{res}{result of \code{mroast} or \code{fry}.}

\item{index}{index of pathways in G}

\item{stats.tab}{a table of feature (e.g. gene) statistics that the Excel file can link to}

\item{n.toptabs}{number of gene set toptables to write to CSV and link to from Excel.}
}
\description{
Write Excel xlsx file with links to CSVs
}
