% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl_pwys.R
\name{xl_pwys}
\alias{xl_pwys}
\title{Make XL data frame for \code{writexl}}
\usage{
xl_pwys(pwy.tab, pwy.nchar = 199)
}
\arguments{
\item{pwy.tab}{A table of pathway results.}

\item{pwy.nchar}{Numeric maximum number of characters allowed in pathway name.}
}
\value{
A data frame with xl links
\describe{
  \item{\code{xl}}{}
  \item{\code{pwy.csv.nms}}{vector of pathway names to write to csv files}
}
}
\description{
Make pathways table into XL data frame for \code{writexl} by adding hyperlink column with cleaned filenames for URLs.
}
