% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zcols.R
\name{add_zcols}
\alias{add_zcols}
\title{Add column to toptab with z-stats}
\usage{
add_zcols(toptab, stat.suffix = "t", p.suffix = "p")
}
\arguments{
\item{toptab}{Table with genes as rows and statistics as columns. Must include a column with t-statistics that
is \code{"t"} or ends with \code{".t"} & a column of p-values is \code{"p"} or ends with \code{".p"}.
Columns with t-statistics and those with p-values must have same prefixes.}

\item{stat.suffix}{SUffix of numeric signed statistics column(s).}

\item{p.suffix}{SUffix of numeric p-value column(s).}
}
\description{
Add column to toptab with z-stats.
}
