% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pvalues.R
\name{combine_pvalues}
\alias{combine_pvalues}
\title{Combine p-values of a feature over multiple p-value columns of an object}
\usage{
combine_pvalues(tab, p.cols = "p|PValue",
  stat.cols = "logFC|slope|cor|rho|Direction", only.p = TRUE,
  alternative = c("two.sided", "greater", "less", "Up", "Down"))
}
\arguments{
\item{tab}{Matrix-like object with statistical columns, some containing p-values.}

\item{p.cols}{Indices or \code{\link{regexp}} with column names or column names suffix of numeric p-value columns.}

\item{stat.cols}{Indices or \code{\link{regexp}} with column names or column names suffix with numeric signed 
statistics, or with \code{"Up", "Down"} values. These should match \code{p.cols}.}

\item{only.p}{Logical; should only combined p-values be returned? If not, returns matrix with z-scores and FDRs also.}

\item{alternative}{Direction of change: \code{"two.sided"}; \code{"greater"} or \code{"less"}, or their synonyms
\code{"Up"} or \code{"Down"}.}
}
\value{
Vector of p-values.
}
\description{
Combine p-values of a feature over multiple p-value columns of an object. If \code{alternative!="two.sided"},
uses the direction of \code{stat.cols} to transform \code{p.cols} into one-sided p-values; this assumes that
these p-values were originally calculated from two-sided tests, as is done in \pkg{limma}.
}
\details{
Z-transform method is used to combine p-values across rows, equivalently to using unweighted 
\code{method="z.transform"} in \code{survcomp::combine.test}.

\code{stat.cols} are ignored if \code{alternative} is \code{"two.sided"}.
}
\examples{
 tab <- data.frame(foo.p=(1:9)/9, bar.p=(9:1)/9)
 combine_pvalues(tab)
}
