% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_signif.R
\name{df_signif}
\alias{df_signif}
\title{Apply \code{signif} function to numeric columns of data frames}
\usage{
df_signif(df, digits = 3)
}
\arguments{
\item{df}{A data frame}

\item{digits}{number of digits passed to \code{signif}.}
}
\description{
Apply \code{signif} function to numeric columns of data frames
}
