% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezcamerapr.R
\name{ezcamerapr}
\alias{ezcamerapr}
\title{A wrapper function for \code{limma::cameraPR} with output to Excel}
\usage{
ezcamerapr(gstats, G, feat.tab, name = NA, adjust.method = "BH",
  alternative = c("two.sided", "greater", "less", "Up", "Down"),
  min.nfeats = 3, max.nfeats = 1000)
}
\arguments{
\item{gstats}{A nmatrix-like data object with rownames & one column of genewise statistic (e.g. z-scores, t-statistics) 
by which genes can be ranked. 
The names should be the same as the rownames of  *feat.tab*}

\item{G}{Gene set list as returned from \code{\link{read_gmt}}.}

\item{feat.tab}{Table of feature (e.g. gene) statistics that the Excel table can link to.}

\item{name}{Name for the folder and Excel file that get written. Set to \code{NA} to avoid writing output.}

\item{adjust.method}{Method used to adjust the p-values for multiple testing. Only for \code{mroast}.}

\item{alternative}{Alternative hypothesis; must be one of\code{"two.sided"}; \code{"greater"} or \code{"less"},
or their synonyms  \code{"Up"} or \code{"Down"}.}

\item{min.nfeats}{Minimum number of features (e.g. genes) needed in a gene set for testing.}

\item{max.nfeats}{Maximum number of features (e.g. genes) needed in a gene set for testing.}
}
\value{
Data frame of gene set statistics.
}
\description{
Test whether a set of genes is highly ranked relative to other genes in terms of differential expression, 
accounting for inter-gene correlation \code{\link[limma]{camera}}.
To find pathways whose genes have large magnitude changes, independent of direction of their change, 
provide \code{abs} of gene stats, and set \code{alternative="Up"}.
It returns a data frame with statistics per gene set, and writes this to an Excel file. 
The Excel file links to CSV files, which contain statistics per gene set.
}
\details{
Pathway (i.e. gene set) names are altered to be valid filenames in Windows and Linux. Numeric columns are
rounded to 3 significant figures.
}
