% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eztoptab.R
\name{eztoptab}
\alias{eztoptab}
\title{Wrapper for limma topTable function}
\usage{
eztoptab(fit, cols = c("P.Value", "adj.P.Val", "logFC"),
  adjust.method = "BH", prefix = NULL, coef = NULL)
}
\arguments{
\item{fit}{Object of class \code{MArrayLM} as produced by \code{\link[limma]{lmFit}} then \code{\link[limma]{eBayes}}.}

\item{cols}{Columns of \code{topTable} output to include. Possibilities include 
\code{"logFC", "AveExpr", "z", "t", "P.Value", "adj.P.Val", "B"}. Some of these column names are then changed here. 
If \code{logFC} is specified, \code{FC} will automatically also be given.}

\item{adjust.method}{Method used to adjust the p-values for multiple testing. Options, in increasing conservatism, 
include \code{"none"}, \code{"BH"}, \code{"BY"}, and \code{"holm"}. See \code{\link[stats]{p.adjust}} for the complete
list of options. A \code{NULL} value will result in the default adjustment method, which is \code{"BH"}.}

\item{prefix}{Character string to add to beginning of column names. \code{NULL} does not add a prefix.}

\item{coef}{Column index or column name specifying which coefficient or contrast of the linear model to test. 
If \code{length(coef)>1}, an F-test will be performed, and \code{logFC} & \code{FC} will not be returned.}
}
\value{
Data frame.
}
\description{
Wrapper for \pkg{limma} \code{\link[limma]{topTable}} function that subsets and changes colnames
}
\details{
If \code{length(coef)>=2}, \code{"z"} will not be returned.
}
\seealso{
\code{\link[limma]{topTable}}.
}
