% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_index.R
\name{g_index}
\alias{g_index}
\title{Get indices of pathways that are the right size}
\usage{
g_index(G, object, min.ngenes, max.ngenes)
}
\arguments{
\item{G}{a gene set list as returned from \code{\link{read_gmt}}.}

\item{object}{A matrix-like data object containing log-ratios or
log-expression values for a series of arrays, with rows corresponding to
genes and columns to samples.}

\item{min.ngenes}{minimum number of genes needed in a gene set for testing.}

\item{max.ngenes}{maximum number of genes needed in a gene set for testing.}
}
\value{
Index of acceptable gene sets in \code{G}.
}
\description{
Get indices of pathways that are the right size.
}
