% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitman.R
\name{hitman}
\alias{hitman}
\title{High-throughput mediation analysis}
\usage{
hitman(E, M, Y, covariates = NULL, verbose = FALSE)
}
\arguments{
\item{E}{A vector representing exposure; can be numeric or a character/factor with nominal groups.}

\item{M}{A numeric matrix-like data object with one row per feature and one column per sample representing the mediator.
Must have more than one feature.}

\item{Y}{A numeric vector of \code{length(E)} of outcomes.}

\item{covariates}{Numeric vector or matrix of covariates.}

\item{verbose}{Logical indicating if messages should be reported.}
}
\value{
Data frame with columns
\describe{
\item{EMY.p}{Overall p-value for mediation}
\item{EMY.FDR}{Overall FDR for mediation}
\item{EM_dir.p}{p-value for E-->M accounting for direction of mediation}
\item{MY_dir.p}{p-value for M-->Y accounting for direction of mediation}
\item{EM.p}{p-value for E-->M, not accounting for direction}
\item{EM.t or EM.F}{F-statistic or t-statistic for E-->M, not accounting for direction}
\item{MY.p}{p-value for M-->Y, not accounting for direction}
\item{MY.slope}{slope of regression for M-->Y, not accounting for direction}
}
and annotation.
}
\description{
High-throughput mediation analysis to test if rows of \code{M} mediate the effect of exposure \code{E} on outcome
\code{Y}.
}
\details{
If \code{E} and \code{Y} have names, and \code{M} has colnames, they should all match. \code{E} and \code{Y}
cannot have \code{NA}s.
}
