% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitman.R
\name{hitman}
\alias{hitman}
\title{High-throughput mediation analysis}
\usage{
hitman(E, M, Y, covariates = NULL, check.names = TRUE)
}
\arguments{
\item{E}{A numeric vector of exposures.}

\item{M}{A numeric matrix-like data object with one row per feature and one column per sample of mediators.
Must have more than one feature.}

\item{Y}{A numeric vector of \code{length(E)} of outcomes.}

\item{covariates}{Numeric matrix-like object with rows corresponding to samples and columns to covariates
to be adjusted for.}

\item{check.names}{Logical; should \code{names(E)==colnames(M) & colnames(M)==names(Y)} be checked?}
}
\value{
Data frame with columns
\describe{
\item{EMY.p}{Overall p-value for mediation}
\item{EMY.FDR}{Overall FDR for mediation}
\item{EM_dir.p}{p-value for E-->M accounting for direction of mediation}
\item{MY_dir.p}{p-value for M-->Y accounting for direction of mediation}
\item{EM.t}{t-statistic for E-->M, not accounting for direction}
\item{EM.p}{p-value for E-->M, not accounting for direction}
\item{MY.t}{t-statistic for M-->Y, not accounting for direction}
\item{MY.p}{p-value for M-->Y, not accounting for direction}
}
}
\description{
High-throughput mediation analysis to test if rows of \code{M} mediate the effect of exposure \code{E} on outcome
\code{Y}.
}
\details{
\code{E} and \code{Y} cannot have \code{NA}s.
}
