% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_contrasts.R
\name{limma_contrasts}
\alias{limma_contrasts}
\title{Apply lmFit, contrasts.fit, & eBayes to one or more contrasts}
\usage{
limma_contrasts(object, grp = NULL, contrast.v, design = NULL,
  weights = NULL, trend = FALSE, block = NULL, correlation = NULL,
  adjust.method = "BH", add.means = TRUE, treat.lfc = NULL,
  cols = c("P.Value", "adj.P.Val", "logFC"))
}
\arguments{
\item{object}{A matrix-like data object containing log-ratios or 
log-expression values for a series of arrays, with rows corresponding to 
genes and columns to samples.}

\item{grp}{Vector of phenotype groups of the samples, which represent valid 
variable names in R. Should be same length as \code{ncol(object)}. If the 
vector is named, names should match \code{colnames(object)}.}

\item{contrast.v}{A named vector of contrasts for \code{\link[limma]{makeContrasts}}.}

\item{design}{the design matrix of the microarray experiment, with rows 
corresponding to arrays and columns to coefficients to be estimated.}

\item{weights}{non-negative observation weights. Can be a numeric matrix of 
individual weights, of same size as the object expression matrix, or a 
numeric vector of array weights with length equal to \code{ncol} of the 
expression matrix, or a numeric vector of gene weights with length equal to
\code{nrow} of the expression matrix.}

\item{trend}{logical, should an intensity-trend be allowed for the prior 
variance? Default is that the prior variance is constant.}

\item{block}{vector or factor specifying a blocking variable on the arrays. 
Has length equal to the number of arrays.}

\item{correlation}{the inter-duplicate or inter-technical replicate 
correlation.}

\item{adjust.method}{method used to adjust the p-values for multiple testing.}

\item{add.means}{logical indicating if (unweighted) group means per row
should be added to the output.}

\item{treat.lfc}{Vector of logFC passed to \code{\link[limma]{treat}} \code{lfc}. It is recycled as needed to match
rows of \code{object}. If given, \code{length(contrast.v)} must be 1. McCarthy & Smyth suggest a 10% fold-change,
which is \code{treat.lfc=log2(1.1)}.}

\item{cols}{columns of \code{topTable} the user would like in the 
output. Possibilities include \code{'logFC', 'AveExpr', 't', 'P.Value', 'adj.P.Val', 'B'}. 
Once selected, the column names of the output may be different than \code{cols}.
If \code{logFC} is specified, \code{FC} will automatically also be given.}
}
\value{
Data frame.
}
\description{
Apply lmFit, contrasts.fit, & eBayes to one or more contrasts by passing each 
object row, phenotype groups, and contrasts to only this function.
}
\details{
If \code{design} is \code{NULL} and \code{grp} is given, design will
  be calculated as \code{model.matrix(~0+grp)}. However, \code{grp} isn't
  needed if \code{design} is provided & \code{add.means} is \code{FALSE}. See
  further details in \code{\link[limma]{lmFit}}.
}
\references{
McCarthy DJ & Smyth GK (2009). Testing significance relative to a fold-change threshold is a TREAT. Bioinformatics 25, 765-771.
}
\seealso{
\code{\link[ezlimma]{limma_cor}}, \code{\link[limma]{lmFit}} and \code{\link[limma]{eBayes}}.
}
