% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_dep.R
\name{limma_dep}
\alias{limma_dep}
\title{Test (conditional) dependence of two variables}
\usage{
limma_dep(object, y, covar = NULL, prefix = "", verbose = FALSE)
}
\arguments{
\item{object}{A matrix-like data object with rows corresponding to features and columns to samples.}

\item{y}{A vector. If not numeric, treated as a nominal group variable. To ensure that \code{object} matches \code{y},
it is checked that \code{colnames(object)==names(y)}. This check will pass if one of these is \code{NULL}.}

\item{covar}{A matrix-like data object with rows corresponding to samples and columns to covariates. If \code{covar}
has rownames, they should match \code{names(y)}.}

\item{prefix}{Character string to add to beginning of column names.}

\item{verbose}{If \code{TRUE} reports messages.}
}
\value{
Dataframe with \code{t-statistic} or \code{F-statistic} and \code{p-value} per row of \code{object}.
}
\description{
Test (conditional) dependence of \code{y} with rows of \code{object}.
}
\details{
This function is not meant to be called directly by the user.
}
