% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_dep.R
\name{limma_dep}
\alias{limma_dep}
\title{Test (conditional) dependence of two variables}
\usage{
limma_dep(object, Y, covariates = NULL, prefix = NULL)
}
\arguments{
\item{object}{A matrix-like data object with rows corresponding to features and columns to samples.}

\item{Y}{A numeric vector or matrix.}

\item{covariates}{A matrix-like data object with rows corresponding to samples and columns to covariates.}

\item{prefix}{Character string to add to beginning of column names.}
}
\value{
Dataframe with \code{t-statistic} or \code{F-statistic} and \code{p-value} per row of \code{object}.
}
\description{
Test (conditional) dependence of \code{Y} with rows of \code{object} via \code{\link[ezlimma]{limmaF}} or 
\code{\link[ezlimma]{limma_cor}}.
}
