% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaF.R
\name{limmaF}
\alias{limmaF}
\title{F-test of each row of object using design matrix}
\usage{
limmaF(object, design, coef = colnames(design), prefix = NULL,
  weights = NULL, trend = FALSE, block = NULL, correlation = NULL,
  adjust.method = "BH", reorder.rows = TRUE, cols = c("F",
  "P.Value"))
}
\arguments{
\item{object}{A matrix-like data object containing log-ratios or 
log-expression values for a series of samples, with rows corresponding to 
genes and columns to samples.}

\item{design}{the design matrix of the experiment, with rows corresponding to 
samples and columns to coefficients to be estimated. Can be used to provide 
covariates.}

\item{coef}{Vector of coefficients to include in F-test. These should be in \code{colnames(design)}. If your design matrix has an intercept, you likely want to exclude it.}

\item{prefix}{character string to add to beginning of column names.}

\item{weights}{non-negative observation weights. Can be a numeric matrix of 
individual weights, of same size as the object expression matrix, or a 
numeric vector of array weights with length equal to \code{ncol} of the 
expression matrix, or a numeric vector of gene weights with length equal to 
\code{nrow} of the expression matrix.}

\item{trend}{logical, should an intensity-trend be allowed for the prior 
variance? Default is that the prior variance is constant.}

\item{block}{vector or factor specifying a blocking variable on the arrays. 
Has length equal to the number of arrays.}

\item{correlation}{the inter-duplicate or inter-technical replicate 
correlation.}

\item{adjust.method}{method used to adjust the p-values for multiple testing.}

\item{reorder.rows}{logical, should rows be reordered by F-statistic from 
\code{\link[limma]{toptable}} or be left in the same order as 
\code{object}?}

\item{cols}{columns of \code{topTable} output the user would like in the 
result. Some column names, such as \code{adj.P.Val} are changed. If \code{logFC}
is specified, \code{FC} will also be given. \code{limma} doesn't return an F-statistic if \code{length(coef)==1}.}
}
\value{
Dataframe.
}
\description{
F-test of each row of object using design matrix for selected coefficients. If there's only one selected coefficient, 
a t-test is applied.
}
\examples{
object <- matrix(rnorm(60), nrow=10, ncol=6)
grp <- rep(c("A", "B", "ctrl"), each=2)
covar <- cbind(covar1=1:6, covar2=rnorm(6))
design <- model.matrix(~0+grp+covar)
lf <- limmaF(object=object, design=design, coef=colnames(design)[1:3])
}
\seealso{
\code{\link[limma]{lmFit}} and \code{\link[limma]{eBayes}}.
}
