% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfc2fc.R
\name{logfc2fc}
\alias{logfc2fc}
\title{Transform a log2 fold-change to a fold-change}
\usage{
logfc2fc(logFC)
}
\arguments{
\item{logFC}{Numeric vector of log2 fold-changes.}
}
\value{
Fold-change.
}
\description{
Transform a log2 fold-change to a (possibly negative) fold-change.
}
\details{
This function converts log fold-changes < 0 to their negative multiplicative reciprocal, e.g. 
fc = 0.5 -> fc = -2, as is commonly done in biology. This function is not meant to be called directly by the user.
}
\examples{
\dontrun{
# not run
logfc2fc(2) #4
logfc2fc(-2) #-4
}
}
