% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_glist.R
\name{map_glist}
\alias{map_glist}
\title{Map pathway list gene IDs to another ID type}
\usage{
map_glist(G, annot, sep.str = " /// ", symbol.col = "Gene.Symbol")
}
\arguments{
\item{G}{Gene set list as returned from \code{\link{read_gmt}}.}

\item{annot}{annotation for the features that has a column of the same type as in gene set list *G*.}

\item{sep.str}{strings that separates symbols if there are multiple symbols for a feature.}

\item{symbol.col}{column name or index for the symbol column in *annot*.}
}
\value{
Gene set list *G* with the symbols replaced by the mapped IDs.
}
\description{
Map pathway list with old gene IDs (e.g. entrez IDs) to another ID type (e.g. gene symbols) using a table with both
annotation types. The original IDs can be separated within a column of annot with a string, e.g. " /// ".
The other IDs should be the rownames of annot. The idea for this is based on the Gene Set Enrichment
Analysis (GSEA) utility [chip2chip](https://www.gsea-msigdb.org/gsea/doc/GSEAUserGuideFrame.html?Chip2Chip_Page).
}
\details{
Gene annotations are transformed to upper-case to avoid missing matches of different cases.
}
