% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mroast_two2one_tailed.R
\name{mroast_two2one_tailed}
\alias{mroast_two2one_tailed}
\title{Convert 2-tailed into 1-tailed p-values from mroast}
\usage{
mroast_two2one_tailed(tab, p.col = "PValue", dir.col = "Direction",
  direction = "Up", nrot = 9999)
}
\arguments{
\item{tab}{Table with statistics from \code{fry}.}

\item{p.col}{Name or index of p-value column.}

\item{dir.col}{Name or index of column giving direction gene set has changed.}

\item{direction}{Direction of gene set change. Can be \code{"Up"} or \code{"Down"}.}

\item{nrot}{Number of rotations used to estimate the p-values for \code{mroast}.}
}
\value{
Vector of p-values.
}
\description{
Convert 2-tailed into 1-tailed p-values from \pkg{limma} function \code{mroast}, accounting for number of 
permutations.
}
