% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiTopTab.R
\name{multiTopTab}
\alias{multiTopTab}
\title{Combine one or more toptables extracted from linear model fit}
\usage{
multiTopTab(fit, cols = c("P.Value", "adj.P.Val", "logFC"),
  adjust.method = "BH")
}
\arguments{
\item{fit}{Object of class \code{MArrayLM} as produced by \code{\link[limma]{lmFit}} then \code{\link[limma]{eBayes}}.}

\item{cols}{Columns of \code{topTable} output to include. Possibilities include 
\code{"logFC", "AveExpr", "z", "t", "P.Value", "adj.P.Val", "B"}. Some of these column names are then changed here. 
If \code{logFC} is specified, \code{FC} will automatically also be given.}

\item{adjust.method}{Method used to adjust the p-values for multiple testing. Options, in increasing conservatism, 
include \code{"none"}, \code{"BH"}, \code{"BY"}, and \code{"holm"}. See \code{\link[stats]{p.adjust}} for the complete
list of options. A \code{NULL} value will result in the default adjustment method, which is \code{"BH"}.}
}
\value{
Data frame
}
\description{
Combine one or more toptables extracted from \code{fit} using \code{\link{eztoptab}}.
}
\seealso{
\code{\link{eztoptab}}.
}
