% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roast_cor.R
\name{roast_cor}
\alias{roast_cor}
\title{Test association of gene sets to phenotype using rotation testing with output
to Excel}
\usage{
roast_cor(object, G, stats.tab, name = NA, phenotype = NULL,
  design = NULL, fun = c("fry", "mroast"), set.statistic = "mean",
  weights = NA, gene.weights = NULL, trend = FALSE, block = NULL,
  correlation = NULL, prefix = NULL, adjust.method = "BH",
  min.ngenes = 3, max.ngenes = 1000, alternative = c("two.sided", "less",
  "greater"), n.toptabs = Inf, nrot = 999, seed = 0)
}
\arguments{
\item{object}{A matrix-like data object containing log-ratios or 
log-expression values for a series of arrays, with rows corresponding to 
genes and columns to samples.}

\item{G}{a gene set list as returned from \link{read_gmt}.}

\item{stats.tab}{a table of feature (e.g. gene) statistics that the Excel 
table can link to.}

\item{name}{a name for the folder and Excel file that get written. Set to \code{NA} 
to avoid writing output.}

\item{phenotype}{Vector of phenotypes of the samples. Should be same length as
\code{ncol(object)}. If the vector is named, names should match 
\code{colnames(object)}.}

\item{design}{the design matrix of the experiment, with rows corresponding to 
arrays and columns to coefficients to be estimated. The first column should be the intercept, 
the second column corresponds to the phenotype, and other columns to covariates. See vignette.}

\item{fun}{function to use, either \code{fry} or \code{mroast}.}

\item{set.statistic}{summary set statistic. Possibilities are \code{"mean"},
\code{"floormean"}, \code{"mean50"}, or \code{"msq"}. Only for \code{mroast}.}

\item{weights}{non-negative precision weights passed to \code{lmFit}. Can be
a numeric matrix of individual weights, of same size as the object 
expression matrix, or a numeric vector of array weights with length equal to
\code{ncol} of the expression matrix, or a numeric vector of gene weights 
with length equal to \code{nrow} of the expression matrix. Set to \code{NULL} to ignore \code{object$weights}. \code{weights=NA} 
 (with length one) doesn't pass weights to \code{limma}.}

\item{gene.weights}{numeric vector of directional (positive or negative) genewise weights. These represent
each gene's contribution to pathways. They are not for precision weights, from \code{weights}. This 
vector must have length equal to \code{nrow(object)}. Only for \code{mroast}.}

\item{trend}{logical, should an intensity-trend be allowed for the prior 
variance? Default is that the prior variance is constant.}

\item{block}{vector or factor specifying a blocking variable on the arrays. 
Has length equal to the number of arrays.}

\item{correlation}{the inter-duplicate or inter-technical replicate 
correlation.}

\item{prefix}{character string to add to beginning of column names.}

\item{adjust.method}{method used to adjust the p-values for multiple testing.}

\item{min.ngenes}{minimum number of genes needed in a gene set for testing.}

\item{max.ngenes}{maximum number of genes needed in a gene set for testing.}

\item{alternative}{indicates the alternative hypothesis and must be one of 
\code{"two.sided"}, \code{"greater"} or \code{"less"}. \code{"greater"} 
corresponds to positive association, \code{"less"} to negative association.}

\item{n.toptabs}{number of gene set toptables to write to CSV and link to from
Excel}

\item{nrot}{number of rotations used to compute the p-values in \code{mroast}.}

\item{seed}{Integer seed to set for reproducility if \code{fun="mroast"}, since \code{mroast} uses random 
simulations. Ignored if \code{fun="fry"}.}
}
\value{
data frame of gene set statistics.
}
\description{
Test association of gene sets to phenotype using rotation testing with
\code{\link[limma]{roast}} using \code{limma} functions \code{mroast} or \code{fry}.
It returns a data frame with statistics per gene set, and writes this to an
Excel file. The Excel file links to CSV files, which contain statistics per
gene set.
}
