% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_xl.R
\name{top_xl}
\alias{top_xl}
\title{Make XL data frame for \code{writexl}}
\usage{
top_xl(pwy.tab)
}
\arguments{
\item{pwy.tab}{A table of pathway results.}
}
\value{
A data frame with xl links
\describe{
  \item{\code{xl}}{}
  \item{\code{pwy.csv.nms}}{vector of pathway names to write to csv files}
}
}
\description{
Make XL data frame for \code{writexl} using top pathways table.
}
