% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two2one_tailed.R
\name{two2one_tailed}
\alias{two2one_tailed}
\title{Convert 2-tailed into 1-tailed p-values}
\usage{
two2one_tailed(tab, stat.col = "logFC|slope|r|Direction",
  p.col = "p|PValue", alternative = c("greater", "less", "Up", "Down"))
}
\arguments{
\item{tab}{Matrix-like object with statistical columns, some containing p-values.}

\item{stat.col}{Index or \code{\link{regexp}} with column name  or column name suffix with numeric signed statistics
or with \code{"Up", "Down"} values.}

\item{p.col}{Index or \code{\link{regexp}} with column name suffix or full column names of p-value columns.}

\item{alternative}{Vector of direction of change: either \code{"greater"} or \code{"less"}. Also permitted are
\code{"Up"} or \code{"Down"}, in which case the \code{stat.col} must include elements \code{"Up"} or \code{"Down"}.}
}
\value{
p-values.
}
\description{
Convert 2-tailed into 1-tailed p-values from a matrix-like object.
}
