% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two2one_tailed.R
\name{two2one_tailed}
\alias{two2one_tailed}
\title{Convert 2-tailed into 1-tailed p-values}
\usage{
two2one_tailed(tab, stat.col = 1, p.col = 2, alternative = c("greater",
  "less"))
}
\arguments{
\item{tab}{table with statistics from \code{fry}.}

\item{stat.col}{name or index of column with signed statistics.}

\item{p.col}{name or index of p-value column.}

\item{alternative}{direction of change. Can be a vector. \code{"greater"} or \code{"less"}.}
}
\value{
p-values.
}
\description{
Convert 2-tailed into 1-tailed p-values from matrix. Alternative "greater" corresponds to positive association, 
"less" to negative association.
}
\details{
This function is not meant to be called directly by the user.
}
