% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two2one_tailed.R
\name{two2one_tailed}
\alias{two2one_tailed}
\title{Convert 2-tailed into 1-tailed p-values}
\usage{
two2one_tailed(tab, p.cols = "p|PValue",
  stat.cols = "logFC|slope|cor|rho|Direction",
  alternative = c("greater", "less", "Up", "Down"), nperm = NULL)
}
\arguments{
\item{tab}{Matrix-like object with statistical columns, some containing p-values.}

\item{p.cols}{Indices or \code{\link{regexp}} with column names or column names suffix of numeric p-value columns.}

\item{stat.cols}{Indices or \code{\link{regexp}} with column names or column names suffix with numeric signed 
statistics, or with \code{"Up", "Down"} values. These should match \code{p.cols}.}

\item{alternative}{Direction of change: either \code{"greater"} or \code{"less"}, or their synonyms
\code{"Up"} or \code{"Down"}.}

\item{nperm}{Number of rotations used to estimate the p-values from \code{\link[limma]{roast}}, or number of 
permutations, where p-value calculated according to Phipson & Smyth (2010). Ignored if \code{NULL}.}
}
\value{
matrix of p-values.
}
\description{
Convert 2-tailed into 1-tailed p-values, possibly using signs of associated statistics, from a matrix-like object.
}
\references{
Phipson & Smyth (2010). Permutation P-values Should Never Be Zero: Calculating Exact P-values When 
Permutations Are Randomly Drawn. Statistical Applications in Genetics and Molecular Biology, 9(1).
}
