% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_linked_xlsx.R
\name{write_linked_xlsx}
\alias{write_linked_xlsx}
\title{Write Excel XLSX file with links to CSVs}
\usage{
write_linked_xlsx(name, res, index, stats.tab, n.toptabs)
}
\arguments{
\item{name}{A name for the folder and Excel file. If \code{NA}, a directory is not created and no file is written.}

\item{res}{A table of pathway results.}

\item{index}{Named list of Indices of pathways in G to write.}

\item{stats.tab}{A table of feature (e.g. gene) statistics that the Excel file can link to.}

\item{n.toptabs}{Number of pathway toptables to write to CSV and link to from Excel.}
}
\value{
Invisibly, the data frame that's written to Excel.
}
\description{
Create directory and write CSV files with stats from \code{stats.tab} per pathway, and write an Excel XLSX file 
using package \code{writexl} that links to the CSVs.
}
