% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_top_xl.R
\name{write_top_xl}
\alias{write_top_xl}
\title{Write Excel XLSX file with links to CSVs}
\usage{
write_top_xl(pwy.tab, feat.lst, feat.tab, name = NA, n.toptabs = Inf)
}
\arguments{
\item{pwy.tab}{A table of pathway results.}

\item{feat.lst}{Named list, with each element containing names or indices of rows of \code{feat.tab}.}

\item{feat.tab}{Table of feature (e.g. gene) statistics.}

\item{name}{Name for the folder and Excel file that get written. Set to \code{NA} to avoid writing output.}

\item{n.toptabs}{Number of gene set toptables to write to CSV and link to from Excel.}
}
\value{
Invisibly, the data frame that's written to Excel.
}
\description{
Create directory and write CSV files with stats from \code{feat.tab} per pathway, and write an Excel XLSX file 
using package \code{writexl} that links to the CSVs.
}
